♻️  cjs generate via tsup and use own dog food 🐶🥫
