+++ config/semantic-release-config/src/zzz.ts
@@ -1,0 +1,3 @@
+const noIhateThisWhyyyyyyyy = () => {}
+
+export { noIhateThisWhyyyyyyyy }
